replaces data source harbor_elb with harbor_loadbalancer (closes #12)
